updates for local use

+++ R/DropBoxUpdate.R
@@ -17,11 +17,16 @@
 #' getTreeMapData(DropBox_dir,budget=budget,data_treemap="/data_treemap.csv")
 
 
-getTreeMapData<-function(DropBox_dir,budget=input$budget,tabset=input$tabset,data_treemap="/data_treemap.csv"){
+getTreeMapData<-function(DropBox_dir,local_dir,budget=input$budget,tabset=input$tabset,data_treemap="/data_treemap.csv",use_DropBox=T){
 
-  df<-drop_read_csv(paste0(DropBox_dir,budget,data_treemap))
-  colnames(df)[5]<-'#'
-
+  if(use_DropBox==T){
+    df<-drop_read_csv(paste0(DropBox_dir,budget,data_treemap))
+    colnames(df)[5]<-'#'
+  }
+  if(use_DropBox==F){
+    df<-read.csv(paste0(local_dir,budget,data_treemap),header=T)
+  }
+  
   data<-list()
   for(i in 1:nrow(df)){
 
@@ -50,13 +55,24 @@ getTreeMapData<-function(DropBox_dir,budget=input$budget,tabset=input$tabset,dat
 #' getTabData(DropBox_dir,tab=tab,budget=budget)
 
 
-getTabData<-function(DropBox_dir,tab=input$tabset,budget=input$budget_year){
+getTabData<-function(DropBox_dir,local_dir,tab=input$tabset,budget=input$budget_year,use_DropBox=T){
 
-  filepath<-file.path(tempdir(),'tabs.json')
-  drop_download(path=paste0(DropBox_dir,budget,'/config/',tab,"_tabs.json"),local_path=filepath,overwrite=T)
-  json<-jsonlite::fromJSON(filepath)
+  if(use_DropBox==T){
+    
+      filepath<-file.path(tempdir(),'tabs.json')
+      drop_download(path=paste0(DropBox_dir,budget,'/config/',tab,"_tabs.json"),local_path=filepath,overwrite=T)
+      json<-jsonlite::fromJSON(filepath)
+      if(file.exists(filepath))(file.remove(filepath))
+      
+  }
+  
+  if(use_DropBox==F){
+    
+      json<-jsonlite::fromJSON(paste0(local_dir,budget,"/config/",tab,"_tabs.json"))
+      
+  }
   #data<-jsonlite::toJSON(data)
-  if(file.exists(filepath))(file.remove(filepath))
+  
 
   return(json)
 
@@ -74,13 +90,13 @@ getTabData<-function(DropBox_dir,tab=input$tabset,budget=input$budget_year){
 #' @examples
 #' updateDropBoxBudget(session,DropBox_dir,budget=input$budget_year,tab=input$tabset,values=values$SummaryAll)
 
-updateDropBoxBudget<-function(session,DropBox_dir,budget=input$budget_year,tab=input$tabset){
-
+updateDropBoxBudget<-function(session,DropBox_dir,local_dir,budget=input$budget_year,tab=input$tabset,use_DropBox=T){
+      
       #Get the tree data
-      Treedata<-getTreeMapData(DropBox_dir,budget=budget,data_treemap="/data_treemap.csv")
+      Treedata<-getTreeMapData(DropBox_dir,local_dir,budget=budget,data_treemap="/data_treemap.csv",use_DropBox)
 
       #Get the tab data
-      Tabdata<-getTabData(DropBox_dir,tab=tab,budget=budget)
+      Tabdata<-getTabData(DropBox_dir,local_dir,tab=tab,budget=budget,use_DropBox)
 
       data<-list()
       data$TabData<-Tabdata
@@ -91,19 +107,33 @@ updateDropBoxBudget<-function(session,DropBox_dir,budget=input$budget_year,tab=i
       session$sendCustomMessage(type="updatebudget",data)
 
       #Get the summaryall data for this budget - we use this for program details at the bottom level
-      filepath<-file.path(tempdir(),'summaryall.xlsx')
-      drop_download(path=paste0(DropBox_dir,budget,'/summaryall.xlsx'),local_path=filepath,overwrite=T)
+      if(use_DropBox==T){
+        
+          filepath<-file.path(tempdir(),'summaryall.xlsx')
+          drop_download(path=paste0(DropBox_dir,budget,'/summaryall.xlsx'),local_path=filepath,overwrite=T)
+    
+          SummaryAll<-read.xlsx(filepath, sheet = 1, startRow = 1, colNames = TRUE,
+                                     rowNames = FALSE, detectDates = FALSE, skipEmptyRows = TRUE,
+                                     rows = NULL, cols = NULL, check.names = FALSE, namedRegion = NULL)
+          SummaryAll$Program<-gsub(":"," - ",SummaryAll$Program)
+          SummaryAll$Program<-gsub("!",".",SummaryAll$Program)
 
-      SummaryAll<-read.xlsx(filepath, sheet = 1, startRow = 1, colNames = TRUE,
-                                 rowNames = FALSE, detectDates = FALSE, skipEmptyRows = TRUE,
-                                 rows = NULL, cols = NULL, check.names = FALSE, namedRegion = NULL)
-      SummaryAll$Program<-gsub(":"," - ",SummaryAll$Program)
-      SummaryAll$Program<-gsub("!",".",SummaryAll$Program)
+          #remove the files!
+          if(file.exists(filepath))(file.remove(filepath))
+      }
+      
+       if(use_DropBox==F){
+        
+          filepath<-file.path(local_dir,budget,'summaryall.xlsx')
+          
+          SummaryAll<-read.xlsx(filepath, sheet = 1, startRow = 1, colNames = TRUE,
+                                     rowNames = FALSE, detectDates = FALSE, skipEmptyRows = TRUE,
+                                     rows = NULL, cols = NULL, check.names = FALSE, namedRegion = NULL)
+          SummaryAll$Program<-gsub(":"," - ",SummaryAll$Program)
+          SummaryAll$Program<-gsub("!",".",SummaryAll$Program)
 
-
-
-      #remove the files!
-      if(file.exists(filepath))(file.remove(filepath))
+      }
+      
 
       return(SummaryAll)
